+++ sample/microtransaction-api/CreateMicroTXFromPrivateDocsSample.php
@@ -5,7 +5,7 @@
 
 require __DIR__ . '/../bootstrap.php';
 
-use BlockCypher\Api\MicroTX;
+use BlockCypher\Client\MicroTXClient;
 use BlockCypher\Auth\SimpleTokenCredential;
 use BlockCypher\Rest\ApiContext;
 
@@ -15,7 +15,7 @@ $apiContext = ApiContext::create(
     array('log.LogEnabled' => true, 'log.FileName' => 'BlockCypher.log', 'log.LogLevel' => 'DEBUG')
 );
 
-$microTX = MicroTX::fromPrivate(
+$microTX = MicroTXClient::sendWithPrivateKey(
     "2c2cc015519b79782bd9c5af66f442e808f573714e3c4dc6df7d79c183963cff", // private key
     "C4MYFr4EAdqEeUKxTnPUF3d3whWcPMz1Fi", // to address
     10000, // value (satoshis)